services/trading/api-service/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0
pyotp==2.9.0
bcrypt==4.1.2
email-validator==2.2.0

# Rate Limiting & Middleware
slowapi==0.1.9

# Caching & Queue
redis==5.0.1
celery==5.3.4

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Crypto & Utils
cryptography==41.0.7
httpx==0.25.2
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0